requirements.txt
PyQt5
pymysql
hyperlpr
Pillow
Keras
Theano
Numpy
Scipy
opencv-python
freetype-py